전체 다 리펙토링

+++ src/main/java/com/cos/blog/repository/CommentRepository.java
@@ -4,10 +4,9 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.cos.blog.model.Comment;
-import com.cos.blog.model.Place;
-import com.cos.blog.model.Report;
+import com.cos.blog.entity.Comment;
+import com.cos.blog.entity.Place;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
 	List<Comment> findByPlace(Place place);
 }
