+++ src/main/java/com/cos/blog/repository/StarRatingRepository.java
@@ -4,12 +4,11 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.cos.blog.model.Comment;
-import com.cos.blog.model.Place;
-import com.cos.blog.model.Report;
-import com.cos.blog.model.StarRating;
+import com.cos.blog.entity.Place;
+import com.cos.blog.entity.StarRating;
 
-public interface StarRatingRepository  extends JpaRepository<StarRating, Integer>  {
+public interface StarRatingRepository  extends JpaRepository<StarRating, Long>  {
+	
 	List<StarRating> findByPlace(Place place);
 
 }
